chore(mime): create mime database used by Gdk.Pixbuf
Gtk depends on the mime.cache file to understand the different images types.
Normally this file leaves in /usr/share/mime/mime.cache, but because we change the variable XDG_DATA_DIRS, we need to have this file inside the repository.
The task *install* calls *update-mime-database* that generates the mime.cache using the freedesktop.org.xml.
Other files are also generated but we don't need them.
The freedesktop.org.xml only have information about the two types that we used, png and svg.